requirements.txt
ffmpeg-python>=0.2.0
Send2Trash>=1.5.0